requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-openai
openai
huggingface_hub
python-dotenv
streamlit